+++ src/roq/cme/gateway.cpp
@@ -188,7 +188,7 @@ void Gateway::operator()(Trace<MarketByPriceUpdate> const &event, bool is_last) 
 
 void Gateway::operator()(Trace<MarketByOrderUpdate> const &event, bool is_last) {
   auto callback = []([[maybe_unused]] auto &market_by_price) {};
-  dispatcher_(event, is_last, mbo_bids_, mbo_asks_, callback);
+  dispatcher_(event, is_last, orders_, callback);
 }
 
 void Gateway::operator()(Trace<TradeSummary> const &event, bool is_last) {