+++ src/roq/cme/tools/security.hpp
@@ -33,7 +33,7 @@ struct Security final {
   struct {
     uint32_t no_chunks = {};
     uint32_t last_chunk = {};
-    std::vector<MBOUpdate> bids, asks;
+    std::vector<MBOUpdate> orders;
     core::mbo::Sequencer sequencer;
     uint32_t resubscribe = {};  // XXX move to sequencer
   } mbo = {};
@@ -43,18 +43,17 @@ struct Security final {
     auto reset = [&]() {
       mbo.no_chunks = {};
       mbo.last_chunk = {};
-      mbo.bids.clear();
-      mbo.asks.clear();
+      mbo.orders.clear();
     };
     if (mbo.no_chunks) {
       if (current_chunk == (mbo.last_chunk + uint32_t{1})) {
         if (current_chunk == no_chunks) {
-          callback(mbo.bids, mbo.asks, true);
+          callback(mbo.orders, true);
           reset();
           return true;
         } else {
           mbo.last_chunk = current_chunk;
-          callback(mbo.bids, mbo.asks, false);
+          callback(mbo.orders, false);
         }
       } else {
         reset();
@@ -62,7 +61,7 @@ struct Security final {
     } else if (current_chunk == uint32_t{1}) {
       mbo.no_chunks = no_chunks;
       mbo.last_chunk = current_chunk;
-      callback(mbo.bids, mbo.asks, false);
+      callback(mbo.orders, false);
     }
     return false;
   }